Noissue - Updated buildtools dependency to fix inconsistent package-lock.json
When ensure_depdencies.py is called from inside adblockpluschrome,
which in return calls "npm install" in both adblockpluschrome/buildtools/
and adblockpluschrome/adblockplusui/buildtools/ the package-lock.json gets
re-generated with entries in a different order, which then causes
ensure_depdencies.py to fail because there are local modifcations.

The order of the entries in buildtools/package-lock.json has been fixed in
https://hg.adblockplus.org/buildtools/rev/2ee884010cf5, but adblockplusui
still referred to a version of buildtools before that change.